refactor: Apply 파일구조 수정 (#239)
