+++ frontend/src/components/Apply/DetailFilter/presenter.jsx
@@ -1,75 +1,26 @@
-import React, { useState } from "react";
+import React from "react";
 import Slider from "rc-slider";
 import "../range.css";
 import CloseIcon from "@mui/icons-material/Close";
-import moment from "moment";
 import { CSSTransition } from "react-transition-group";
-import PropTypes, { number } from "prop-types";
 
 import * as S from "../style";
 
 const { createSliderWithTooltip } = Slider;
 const Range = createSliderWithTooltip(Slider.Range);
 
-DetailFilter.propTypes = {
-  filter: PropTypes.shape({
-    schoolId: PropTypes.number,
-    schoolName: PropTypes.string,
-    paymentType: PropTypes.string,
-    minPay: PropTypes.string,
-    doTime: PropTypes.string,
-    startAndEnd: PropTypes.arrayOf(number),
-    startTime: PropTypes.string,
-    endTime: PropTypes.string,
-  }),
-  setFilter: PropTypes.func.isRequired,
-  setFilterSwitch: PropTypes.func.isRequired,
-  filterReset: PropTypes.func.isRequired,
-};
-
-export default function DetailFilter({
-  filter,
-  setFilter,
-  setFilterSwitch,
+export default function Presenter({
   filterReset,
+  setFilterSwitch,
+  paymentType,
+  filter,
+  changeFilter,
+  minPay,
+  doTime,
+  timeArr,
+  setFilter,
   search,
 }) {
-  const createTimeArr = () => {
-    var timeArr = ["00:00"];
-    var time = moment().set("hour", 0).set("minute", 0).set("second", 0);
-    for (var i = 0; i < 24; i++) {
-      timeArr.push(time.add(1, "hours").format("HH:mm"));
-    }
-    return timeArr;
-  };
-  const [timeArr] = useState(createTimeArr);
-
-  const changeFilter = (e) => {
-    const copy = { ...filter };
-    copy[`${e.target.name}`] = e.target.value;
-    setFilter(copy);
-  };
-
-  const paymentType = [
-    { name: "전체", value: "" },
-    { name: "현금", value: "CACHE" },
-    { name: "물품", value: "GOODS" },
-  ];
-  const minPay = [
-    { name: "전체", value: "" },
-    { name: "1만원 미만", value: "9999" },
-    { name: "1만원 이상", value: "10000" },
-    { name: "3만원 이상", value: "30000" },
-    { name: "5만원 이상", value: "50000" },
-  ];
-  const doTime = [
-    { name: "전체", value: "" },
-    { name: "30분 이내", value: "30" },
-    { name: "1시간 이내", value: "60" },
-    { name: "3시간 이내", value: "180" },
-    { name: "3시간 이상", value: "181" },
-  ];
-
   return (
     <S.FilterBox>
       <div
