+++ frontend/src/components/Apply/SearchBar/presenter.jsx
@@ -1,42 +1,20 @@
 import React from "react";
 import DatePicker from "react-datepicker";
 import "../date-picker.css";
-import PropTypes, { number } from "prop-types";
 import { CSSTransition } from "react-transition-group";
 
 import SchoolSelect from "../../Common/SchoolSelect";
 
 import * as S from "../style";
-import { useRecoilValue } from "recoil";
-import { userState } from "../../../recoil/userState";
-
-PSearchFilter.propTypes = {
-  search: PropTypes.func.isRequired,
-  filter: PropTypes.shape({
-    schoolId: PropTypes.number,
-    schoolName: PropTypes.string,
-    paymentType: PropTypes.string,
-    minPay: PropTypes.string,
-    doTime: PropTypes.string,
-    startAndEnd: PropTypes.arrayOf(number),
-    startTime: PropTypes.string,
-    endTime: PropTypes.string,
-  }),
-  setFilter: PropTypes.func.isRequired,
-  date: PropTypes.object.isRequired,
-  setDate: PropTypes.func.isRequired,
-};
-
-export default function PSearchFilter({
+export default function Presenter({
   feedList,
-  search,
+  userSchoolId,
   filter,
   setFilter,
   date,
   setDate,
+  search,
 }) {
-  const userSchoolId = useRecoilValue(userState).schoolId;
-
   return (
     <CSSTransition in={!feedList} classNames="move" timeout={500}>
       <S.SearchBox>
